+++ src/IdentityServer/Infrastructure/Clock/IClock.cs
@@ -2,7 +2,6 @@
 // See LICENSE in the project root for license information.
 
 
-using Microsoft.AspNetCore.Authentication;
 using System;
 
 namespace Duende.IdentityServer;
@@ -16,40 +15,4 @@ public interface IClock
     /// The current UTC date/time.
     /// </summary>
     DateTimeOffset UtcNow { get; }
-}
-
-class DefaultClock : IClock
-{
-    private readonly TimeProvider _timeProvider;
-
-    public DefaultClock()
-    {
-        _timeProvider = TimeProvider.System;
-    }
-
-    public DefaultClock(TimeProvider timeProvider)
-    {
-        _timeProvider = timeProvider;
-    }
-
-    public DateTimeOffset UtcNow { get => _timeProvider.GetUtcNow(); }
-}
-
-class LegacyClock : IClock
-{
-#pragma warning disable CS0618 // Type or member is obsolete
-    private readonly ISystemClock _clock;
-#pragma warning restore CS0618 // Type or member is obsolete
-
-#pragma warning disable CS0618 // Type or member is obsolete
-    public LegacyClock(ISystemClock clock)
-#pragma warning restore CS0618 // Type or member is obsolete
-    {
-        _clock = clock;
-    }
-
-    public DateTimeOffset UtcNow
-    {
-        get => _clock.UtcNow;
-    }
 }
